// src/app/_dto/ibr.dto.ts
export class IbrDto {

    coCode: string;

    dvCode: string;

    brCode: string;

    brName: string;

    brShrtName: string;

    entrySDt: Date;

    entryEDt: Date;

    add1: string | null;

    add2: string | null;

    add3: string | null;

    add4: string | null;

    tel: string | null;

    fax: string | null;

    email: string | null;

    website: string | null;

    commisionAc: string | null;

    discountAc: string | null;

    brNameAr: string | null;

    add1Ar: string | null;

    add2Ar: string | null;

    add3Ar: string | null;

    add4Ar: string | null;


}