// src/app/_dto/lov/iac4-lov.dto.ts
/*
    Created By  : Arun Joy
    Created On  : 06-01-2020
    Created For : For creating dto of Iac4Lov
*/
export class Iac4LovDto{
    
    coCode: string;

    acCode: string;

    acName:string;

    curCode?: string | null;

    add1: string | null;

    add2: string | null;

    add3: string | null;

    add4: string | null;

    tel: string | null;

    fax: string | null;

    contact1: string | null;
 
    email: string | null;

    website: string | null;

    crPeriod: number;

    expGrpCode: string;

}